Initial commit of the generic SerializedScriptValue.
Now the SerializedScriptValue uses SerializationDelegate and
ScriptValue of the language-specific types. It serializes to and from
JavaScript and Ruby using the same structured clone algorithm. Since
the algorithm is the same, the serialized values can be transferred
between languages. A JavaScript JSON object will turn into a Ruby hash.
Any properties on objects will be stored in the hash as symbols.